requirements.txt
Django==1.4.3
distribute==0.6.27
dj-database-url==0.2.1
psycopg2==2.4.5
wsgiref==0.1.2
gunicorn==0.14.6
south
eveapi==1.2.6
pyzmq
mysql-python